Delete unnecessary manual dispatch table for empty.SymInt
The fastpath here handled the situation when Python unconditionally
dispatched to SymInt overloads.  Now that Python doesn't do that, it
is less important for this to be fastpath-ed.  (The only reason you
might still want to fastpath is to optimize calls to this function
from composite kernels which are written only once with SymInt, but
I think it's not worth it for this case.)

Signed-off-by: Edward Z. Yang <ezyangfb.com>

ghstack-source-id: 52d1c06ebe8d9f2f2c9aace7c9b184e6812b85af
Pull Request resolved: https://github.com/pytorch/pytorch/pull/83592